update to spack.yamls to include specific py ver
The fact that the desired python does not contain a hash
in the recipe seems to introduce a bug in the concretization.
This bug results in a version of python other than what
is requested being built. This bug is fixed by explicitly
having the python version in the environment.