Update to environment spack yamls to fix python
The appplications spack.yaml required an update to ensure that
python was the appropriate version. This required spliting list
into those that directly depend on python or depend on plumed that
depends on python from those that do not have a dependency.

Also commented out the amd math library build.