Added script to manage user installations of software
If users use spack to install software, after updates
the users may have software that needs to be reinstalled
as the compilers and older dependencies are no longer
present in the currenty spack deployment. This script
will find all packages given a compiler@version and
uninstall and reinstall as required. Further updates
and ammends to the script are likely.