Re-arrange repository structure
Scripts, fixes and package recipes should be independent of the Pawsey
system spack is installed on. Moving these directories to the root level
will avoid code replication in the future.

On the other hand we will have system specific configuration files.
These are stored under systems/<system-name>.

The examples directory has also been removed as not applicable anymore
(we may want to add a tutorial at some point in the future?)