Update to externals in packages.yaml
Added a few specific libraries that are now part of libfabric dependencies, including nominally packages that are easily built with spack but could have a version difference with that pulled in by libfabric. This setup has not been tested with a full build but with a few key packages like netcdf-c, petsc.

Also minor reorganisation of where packages are listed and some comments.